Migrate from prototype code
