+++ src/sode/python/sode/soundcloud/playlist.py
@@ -1,9 +1,9 @@
 from requests import Response
 
-from sode.shared.oauth import AccessToken1
+from sode.shared.oauth import AccessToken
 
 
-def any(access_token: AccessToken1, user_id: str) -> Response:
+def any(access_token: AccessToken, user_id: str) -> Response:
     """Fetch any playlist for that user"""
 
     # https://developers.soundcloud.com/docs/api/explorer/open-api
